requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0

python-dotenv>=1.0.0
python-multipart>=0.0.9

pydantic>=2.10
pydantic-ai>=1.0.0
anthropic>=0.45.0,<1.0.0



